Avoid using authenticated URL post-logout
